ReplicatedPG.cc: fix snap_trimmer object context error
Previously, snap_trimmer would get the clone object information from the
object store rather than using find_object_context.  This would cause
the cached version to not be updated with the new version in the case
that the object information got updated.  As a result, the need field of
the missing object could get a stale version inconsistent with the most
recent logged version.

Signed-off-by: Samuel Just <samuel.just@dreamhost.com>