MDSMap: split up find_standby_for into multiple functions.
Usage of this function is rapidly diverging, in terms of what
is desired.
We now have "find_standby_for", which selects a standby for a
particular MDS (preferring those specifically set to the MDS, then
those asking to standby for somebody); "find_unused_for" which
finds an unused MDS which does NOT want to be a standby-replay
or a standby for anybody in particular; and "find_replacement_for",
which takes any MDS it can get but prefers those which have been
shadowing the given MDS or standing by for it specifically.

Since we've demoted standby_for_name to a third-class check, don't use
it when considering whether an mds is unused.
This field is now used to pair MDSes, not to bind them to particular
ranks. So the "primary" of the pairing will have standby_for_name
set but still expects to be assigned a rank.

Also, of course, switch the MSDMonitor to call the correct function
in the correct places.

Signed-off-by: Greg Farnum <gregory.farnum@dreamhost.com>