MDSMonitor: Don't create new map for standby-replay spam.
If an MDS is unable to get into the standby-replay state for some
reason (MDS it should be following doesn't exist yet, there aren't
any open MDSes, etc) it will spam the Monitor with beacons asking
to change state. These will always go to prepare_beacon since
they're asking for a state change, but can't be granted.
When this happens, return false, not true!

Signed-off-by: Greg Farnum <gregory.farnum@dreamhost.com>