remove ancient mds_local_osd option
This used to add a few osds to the end of the osdmap, not included in the
crush map, with the assumption that they would be colocated with the MDSs
and used as the primary MDS journal copy.  Not a useful hack anymore; you
can do this all explicitly by setting up your own crush map.

Signed-off-by: Sage Weil <sage.weil@dreamhost.com>