osd: verify object version during push
Fail to push if the ondisk version doesn't match the version we want to
send.

This isn't supposed to happen. If it does it means we have a bug somewhere
else.  Log something to the error log and don't push.  This is better than
the current behavior, which goes into a loop (repeatedly pulling the object
and retrying when it's not the right version).

Signed-off-by: Sage Weil <sage.weil@dreamhost.com>