CDir: Don't write out the header on a partial commit.
If we write out the header as part of a partial commit, and then
fail to complete a subsequent commit (network error, we crash, etc)
then the on-disk version of the directory is not correctly versioned.
The fact that some dentries are of a newer version than others
is okay because we will fix it up during journal replay, but if
the header says the directory is fully committed to the end of
the journal that won't happen!
So, take advantage of how messages between two daemons are strictly
ordered, and how messages for a given PG are strictly ordered, and
simply include the partial commit that contains the new header last.

Signed-off-by: Greg Farnum <gregory.farnum@dreamhost.com>