osd: fix missing.rm()
The version specifies which version of the object no longer should be
missing.  We should thus remove it from the missing set if we needed
anything less than OR EQUAL to that version.  (If we are missing something
newer, then missing.rm() is a no-op.)

Make the argument name less weird while we're at it.

Reported-by: Henry Chang <henry.cy.chang@gmail.com>
Signed-off-by: Sage Weil <sage.weil@dreamhost.com>