filer: set RMW bit on probe
Setting the RMW bits on the probe stat call will make the OSD wait for
pending writes on the object to flush to disk.  This was a problem for MDS
takeover: the old instance had writes mid-commit on the OSD when the
journal probe/stat came in, and it didn't "see" the size change because
the PG was in 'delayed' mode and the old MDS's write hadn't been queued
for disk locally.  Simply claiming we are a RMW op will force the PG
into rmw mode so that we know the prior write will be visible.  This is
just fine for the metadata workload.

Fixes: #805
Signed-off-by: Sage Weil <sage.weil@dreamhost.com>