osd: wait for handle_osd_map transaction ondisk without doing a full sync
Doing a full sync (forcing a btrfs transaction etc) was just wrong here.
All we (might) care about is whether our Objectstore::Transaction is
stable (in journal or fs) or not.

We are still waiting for those operations to flush to the fs (to be
readable).  That may not be necessary either, but shouldn't have a big
performance impact.

Signed-off-by: Sage Weil <sage.weil@dreamhost.com>