osd: factor pg get-or-create code into common helper
handle_pg_notify and _process_pg_info both lookup or create a PG based
on an incoming message.  Factor that code into a common helper.  There
were a few differences in that the pg notify handler code deals with
more cases (namely, pg creation), but this is harmless for the more
general _process_pg_info caller.

Closes: #577
Signed-off-by: Sage Weil <sage.weil@dreamhost.com>