mds: remove bad open_snaprealm()
This was added in b438b3d65b478a25ae1b9cab2cdd16c851d65fc8.  We don't
want it here, though, because this is a _remote_ dentry rename and all
we are doing to the inode is adjusting nlink.  No snaprealms are involved
because the inode isn't moving in the namespace.

The question is what realm we meant to open... :/

Signed-off-by: Sage Weil <sage@newdream.net>