mds: fix client session removal on journal replay
We want to remove the client session from the map as long as it is not
attached to an actual messenger Connection.  This key point got lost
somewhere the last time the session states were restructured.  It is now
explicit.

This fixes the symptom where a recovering MDS reconnect has to time out on
clients that cleanly closed their sessions.

Also, fix a use-after-free when (uselessly) printing the session state.

Signed-off-by: Sage Weil <sage@newdream.net>