common: run_cmd: fix bad implicit conversion
Since NULL is really just a macro defined to be 0, we must use
(char*)NULL or similar to force the compiler to use a true pointer value
as the last argument to the run_cmd varargs function. Otherwise, the 0
gets promoted to an int, which probably is not the same length as a
pointer these days (32 vs. 64.)

Signed-off-by: Colin McCabe <colin.mccabe@dreamhost.com>