osd: process missing when log is empty
There are important cases where the replica will send a missing set and
empty log to the primary during peer (e.g., when the primary asks for it).
In that case, we need to go into pg->proc_replica_log so that peer_missing
gets updated and peering can complete.  It is okay from proc_replica_log's
perspective to pass in an empty log; it will have no effect.  Fix the
if() guard appropriately.

Note that the only path into _process_pg_info where missing is NULL is
the handle_pg_info path, which is used for primary->replica "activate now"
messages, updates after already active, and for replica->primary "ok i
activated" messages.

Signed-off-by: Sage Weil <sage@newdream.net>