ReplicatePG,PG: SnapTrimmer state machine skeleton
Currently, snap_trimmer does not wait for replicas to apply the object
removal repops before updating the info with the removed snapshot and
sending out infos.  Thus, there is a race between the replica applying
the object removal transactions and recieving the info prompting it to
remove its local snap collection.  In the cases where the info beats the
transaction to the file system, we get the infamous -39 ENOTEMPTY error
crashing the OSD.

One solution would be to block in snap_trimmer until we get the
responses.  This would, unfortunately, tie up a disk_tp thread.  Moving
to a state machine would:
a) guarrantee that the snap_trimmer state is cleaned up when the pg
resets (using on_change())
b) remove the necessity of blocking in a disk_tp thread (we currently
unlock and relock the pg lock between operations causing us to block on
another thread using the pg).  Rather than unlocking and relocking, we
would instead unlock and requeue ourselves in the snap_trim_wq while we
wait allowing the thread to do something else.

Signed-off-by: Samuel Just <samuel.just@dreamhost.com>